Changed length of cropped signal from 2*CA.spacing for only CA.spacing in all three Signals
Changed the method of auto-cropping with idea that
time between incident and reflected signals equal to
two times of wave propagation from gage to bar side and
time between incicdent and transmitted signals equal to
time of wave propagating from first gage to second gage.
It is only one problem with specimen's sound velocity,
because we don't know it, so usage the bar's sound velocity
for the specimen is a better than nothing.